# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Split Lease Edge Functions API",
    "version": "2.0.0",
    "description": "Complete API documentation for all Split Lease Supabase Edge Functions.\n\n## Architecture Pattern\nAll endpoints follow the action-based routing pattern:\n```json\n{\n  \"action\": \"action_name\",\n  \"payload\": { /* action-specific data */ }\n}\n```\n\n## Authentication\nMost endpoints require Bearer token authentication via Supabase Auth.\nInclude the header: `Authorization: Bearer <access_token>`\n\n## Response Format\nAll responses follow this structure:\n```json\n{\n  \"success\": true|false,\n  \"data\": { /* response data */ },\n  \"error\": \"Error message if success is false\"\n}\n```\n\n## Day Indexing\nAll day indices use JavaScript's 0-based standard (matching Date.getDay()):\n| Day   | Sun | Mon | Tue | Wed | Thu | Fri | Sat |\n|-------|-----|-----|-----|-----|-----|-----|-----|\n| Index |  0  |  1  |  2  |  3  |  4  |  5  |  6  |\n",
    "contact": {
      "name": "Split Lease Team",
      "email": "splitleaseteam@gmail.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://splitlease-backend-dev.supabase.co/functions/v1",
      "description": "Development"
    },
    {
      "url": "https://splitlease-backend-live.supabase.co/functions/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:54321/functions/v1",
      "description": "Local development"
...
  "paths": {
...
    "/pricing": {
      "post": {
        "summary": "Pricing calculations",
        "description": "Calculate pricing for listings.\n\n**Actions:**\n- `health` - Health check\n",
        "tags": [
          "Listings"
        ],
        "operationId": "pricing",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "examples": {
                "health": {
                  "summary": "Health check",
                  "value": {
                    "action": "health",
                    "payload": {}
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase Auth JWT token"
      }
    },
...
  }
}